Manage page list globally
This reduces duplicated loading effort, but more importantly, allows us to make sure that the previous/next buttons in the detail view iterate through the same list of pages that we are displaying in the list view (when we add sorting/filtering in the future, this will matter more).

Another solution to this might be to make the listing component own the detail view, which we might still want to do if we wind up getting to a point where we display multiple different listing views. If we move to using Redux for state, this is also the sort of data that is well handled by a global state store.