Make WebMonitoringDb into a configurable class
The main instance is constructed by the WebMonitoringUi top-level component and passed down to other components via React context. Context can be a little thorny, so it *may* make sense to wrap this with something nice like Redux's <Provider> in the future or even to have a shared singleton available directly from the WebMonitoringDb module. I think this works well enough for now.