Make first row of both sheets column headers
It was previously inconsistent between the two. This lets us have helpful headers to remind people using the sheets how they should be formatted.

As a side bonus, I realized I didn't need to provide row indices for the end of the range -- if left off, it will include all rows with data!