Sandbox rendered diff iframes
Also test out the idea of retrieving content and modifying it before shoving it into the iframe. In previous commits, we simply loaded the archived page, which fails to get subresources from the right place (and probably some other issues). We can solve that with new API endpoints that process that page and insert things like <base> tags, but this is an experiment with doing it client-side.