Add experimental version timeline to ChangeView
The timeline shows all of a Page's versions in chronological order, with each item sized by how much changed between that version and the preceding version (when that info is available -- only for Versionista data right now). Versions that are flagged as errors have an '!' instead of a circle.

There are likely a number of other useful things we could visualize in this timeline, like whether a version is attached to any changes with annotations, whether a version has been marked as significant, and so on.

For now, this is just an experiment and the implementation needs more than the 2 hours of work here. If people like it, it could replace the dropdowns for selecting versions.