Start removing Typescript, backfill with Babelify
This lets us incrementally convert to JS and test along the way. One downside is that importing .jsx files requires specifying the extension. We should be able to back that out after the full conversion is done, though. (In theory, "allowJs" is supposed to handle this, but it doesn't seem to be working for me.)