Include change versions in URL
Page URLs now look like:

    /page/{id}/{from}..{to}

...which follows a similar scheme as our API (but no `/changes/` in between page ID and change ID). This is kinda important so that specific changes are linkable; people can list them in spreadsheets, chat, etc.

This affects several components because it removes control over the change to be rendered from ChangeView. Now the from and to versions are props. This unfortunately made rendering PageDetails a lot more complicated, but also I hope it's easier to see the various cases that can happen there now.