Include some context in the changes-only diffs
For now, trimmed context is represented by a simple ellipsis. In the future, we might want some more custom rendering.

We trim to 3 lines of context or, if the context is on a single line, trim that line to 300 characters. Those values are for either side of the context, i.e. 3 lines/300 chars after a change and another 3 lines/300 chars before a change.