Split context when trimming
Instead of rewriting the context with an ellipsis somewhere in the text, split the context into multiple contexts. This will allow us to treat the ellipsis (or whatever in the future) as a separate element that we can easily style, make interactive (e.g. click it to expand), and so on.