Clean up left-behind cruft when removing changes
When removing the `<ins>` or `<del>` elements from one side of a change, we often get left with a bunch of empty container elements we don't actually want because our diff is actually a text diff that is sensitive to the tree structure, rather than a diff of the tree structure itself. Take this example diff output:

    <ul>
      <li><del>Some removed text</del></li>
      <li><del>More removed text</del></li>
    </ul>

If we just remove the `<del>` elements, we are left with a bunch of empty bullets. We should (probably) have removed the whole `<ul>`, though (unfortunately, we don't know this for sure, but it's a pretty good guess).

This change attempts to cover the above case by marking the parents of every removed change element and then doing a second pass to see if those parents are now totally empty and, if so, remove *them,* too. This keeps going recursively until a parent is non-empty.