Show error messages when diffs do not load
In DiffView, we now handle errors by making the `state.diff` the error instance. When rendering, we check whether the diff is an error object or normal data and render an error message if necessary.

This also changes the message for unimplemented diffs to display similarly, but with yellow "warning" colors.

Finally, this also adds a generic function to the WebMonitoringDb service for rejecting promises as error instances when they contain error data.

Fixes #112