Add "highlighted rendered" diff
The improved HTML differ over on the processing side finally returns pages that are safe to render without first ripping them apart, so we can finally implement a visual diff with all the changes inline together (instead of side-by-side). This is pretty much the same as `SideBySideRendered`, but only has one frame and doesn't need to do as much processing of the HTML source code.