Babelize with the `env` instead of ES versions
This uses browserslist to apply specifically the transforms needed for the selected browsers (like autoprefixer does), rather than applying all the transforms for given ECMAScript spec versions. The spec version presets are being deprecated with the upcoming Babel v7 release: https://babeljs.io/blog/2017/12/27/nearing-the-7.0-release#deprecated-yearly-presets-eg-babel-preset-es20xx

Note we have duplicated environment configurations between `package.json` and `.babelrc`. There's no reasonable way of avoiding this for now -- Jest compiles files directly with Babel instead of Webpack (compiling with Webpack is possible but involves a *way* more complicated setup; we don't need those features for our Jest tests yet), so it needs the `.babelrc` file around so that Babel can be configured independently of our Webpack file. The version of the `babel-preset-env` package that will be released with Babel v7 can read the `package.json` file to get this data (like autoprefixer already does), so we can remove this once that is released and we upgrade.