Update Dockerfile, upgrade to Latest LTS Node.js
The Dockerfile mainly needed to remove gulp, but it also got more complex here in the name of producing a much smaller image. Because we use Zopfli for compressing the client-side JS, we need Python, GCC, and other build tools available, which add *a lot* of overhead. Now we have a build system with separate stages:

- `dev` builds an image with all development dependencies. You can also use this for working inside of or to run tests.
- `build` builds the production client assets internally.
- `release` is a lightweight image with no extra dependencies that pull the production assets from the `build` image. Because it's the last stage, it's also the default -- so you don't need to target it specially when building.