Re-orient DiffSettings's API around an object
DiffSettings takes a `settings` object that is just a collection of all the various diff settings. It calls `onChange` with an updated object containing new settings whenever the user interacts with it. At current, there's only one setting (`removeFormatting`), but this makes it easy to add new settings for any diff type.