Make sure the page lists are cleared after login
If the user is not logged in, but our routing causes us to try and load pages, the page lists will be populated with errors, which will then cause problems when trying to render later *after* successfully logging in. Now we clear the lists after login, since they necessarily should be dropped and reloaded at that point anyway (you might have logged in as a different user than before, for example).