Remove deprecated zopfli-webpack-plugin
Instead of updating this plugin, the maintainers have finally deprecated it and suggested using the pluggable functionality of compression-webpack-plugin instead. (That's probably because the underlying node-zopfli package seems dead and there are a variety of available replacements -- and they don't wnat to pick one for everybody.)

So this uses node-zopfli-es as the implementation, which is an upgraded version of node-zopfli, using native bindings and the new N-API for Node.js 8+. The other good option here is `@gfx/zopfli`, which is written built with WebAssembly instead of actual native bindings. As a result, it's notably slower (seems to be about 50%-100% slower for our build). On the flip side, it would be easier to build on a variety of systems, so I'm open to changing to it if people want to.