Manage dependencies with NPM instead of Yarn
While NPM is definitely slower and lacks a few of the nicer or fancier features of Yarn, it is more commonly used and comes with Node.js. We don't actually *need* any of those fancy Yarn features (though I will miss them), and switching to it is one less major dependency or tool someone has to install in order to get started. We've generally agreed that's a worthwhile tradeoff, so this switches dependency management to NPM.

Fixes #281.