Add fallback for Intl api #155 (#229)
When we initially added the issue about needed a fallback for `Intl.DateTimeFormat`, the second-latest version of Safari was still reasonably common on Mac and needed support. Since then, we've had another major version of Safari released, and browsers that don't support this feature are comparatively rare. So this simplifies our fallback by not using a full polyfill, and instead just using `Date.toLocaleString()`, a method that's been around for more than a decade and has pretty universal support. It won't have quite the same formatting we want, but it will produce a readable, localized date, and we'll only wind up falling back it rarely.

Doing this means we don't incur any overhead for shipping a full date localization polyfill.

Thanks to @marlungu for the initial work here!