Remove versionista diff link in change view (#374)
The Versionista link was a transitional tool, and analysts are generally no longer using it. We’ve also turned off all the extra paid functionality on our Versionista accounts, and most of the Versionista data we have stored now is aging out of Versionista anyway, so it’s not useful to keep this Versionista diff link around.

Fixes #370.