Switch from binary Zopfli to WebAssembly
We previously used a compiled binary implementation of Zopfli (`node-zopfli-es`) for compressing production builds. It's a lot faster than any other options, but it makes installing and setting up the project a lot harder (see `node-gyp` setup on Windows, for example: https://github.com/nodejs/node-gyp#on-windows). Using a version of Zopfli compiled down to WebAssembly is slower, but it's still not that slow in absolute time. On my machine, production builds now take roughly 10.5 seconds instead of 7.5 seconds. That seems like a worthwhile tradeoff to make installation more-or-less painless.