Document public user credentials on staging DB (#410)
With default settings, the UI uses the staging DB for its data, and that database has a “public access” user with credentials anybody can use. We don’t document that here, though! This adds info about those credentials to the both the `.env.example` file and the setup guide in the README.