Upgrade to React Router v5 (#421)
Update React Router to v5 for compatibility with the latest React (v4 used some method that were deprecated and are going to be removed in the next React release). The major version number bump is because of a packaging issue, and there are no actual backwards incompatibilities.

Also adds @jatinaroragit to contributors list.

Fixes #411.