Narrow the places Jest should scan
I'm seeing a "too many open files" error quite often when running `npm run test-watch`, and I believe I've heard the same from others. I'm not sure what files it's scanning (we really don't have that many!), but it's not `node_modules`. So instead of trying to figure out what to tell it to ignore, I just narrowed the places it should look in the first place.