Update browsers to work with new browserslist
Something seems to have changed in either caniuse-lite or browserslist that is causing it to return results that include a lot more browsers than you'd expect for our query. (Part of this is the odd options babel-preset-env sets for browserslist, but those options didn't change in this update -- it's more how browserslist handles them.) This adds a lot more restrictions to our browsers query so that results are more in line with what we used to have. The changes might look surprising, though:

- Android browser now tracks with Chrome, but the way results are getting treated now is causing us to include a huge amount of actually unnecessary stuff because 'android' is in the results. We're still compatible with recent Android versions by virtue of having the last 3 Chrome versions.
- `not dead` - not sure why we weren't specifying this before, but it cleans up a lot of the noise that was causing problems (like IE Mobile now showing up in the results, but then babel-preset-env transforming it to just IE, and winding up building for IE 9 even when the config says `not ie < 1000` - ugh).

It turns out we also had conflicting browser lists in `package.json` and `.babelrc`. This updates the one in `package.json` (since other tools besides Babel can also use it) and drops the one in `.babelrc`.