Switch back to webpack-cli from webpack-command (#493)
Once upon a time, webpack-command was the new, lighter-weight, future version of webpack-cli. I hadn't been paying close attention, but that's no longer the case, and webpack-command has been spun off into an independent project that is more-or-less in maintenance mode. There's a new fancy light-weight client (webpack-nano), but it is also not official, and I think we're better off to just switch back to the standard webpack-cli.