HOTFIX: Revert css-loader to v3.6.0
css-loader v4+ and babel-plugin-react-css-modules are not compatible with each other. Using them together relies on the fact that they happen to have similar methods for building identifier names for CSS rules, and that similarity disappeared in css-loader v4. Their *might* be a workaround for this by using functions instead of template strings for identifier names, but that requires a little more exploration. For now, we can solve this by reverting css-loader.