Flag versions that were errors in version dropdown (#638)
In the dropdown for selecting versions of a page to compare, we now show an "x" and the status code next to versions that were errors. This is helpful when trying to skip over intermittent errors or determining how long a page has been erroring for.

This display isn’t especially great, and we’d held off on doing this before because we planned to solve this much more nicely in #98 with a more complex and space-efficient timeline-based version selector. However, that’s not actually going to get done anytime in the near future. Meanwhile, analysts are wasting lots of time comparing different versions because they lack this info, so the less-than-ideal solution here is very worthwhile.