Redirect to the correct URL for merged pages (#660)
When loading page data on the page details view, we might get a redirect from the API because a page was merged. This reflects that redirect in the UI's URL, ensuring that old change links redirect to the same view for the new, merged page.

Merged pages are new in edgi-govdata-archiving/web-monitoring-db#793.