Load version subresources from correct URL
We've been using the page's `url` property to determine  the base URL from which to load subresources when showing diffs. However, a page's URL can shift over time, and versions can come from drastically different looking URLs. That means the correct "base" URL for a version depends on the *version,* not the page it's attached to, and also needs to take redirects into account.

For example, a page's URL might have been `https://epa.gov/somewhere`, but one of the versions might have come from `https://epa.gov/somewhere/else`, which itself might have redirected to `https://differentsubdomain.epa.gov/a/whole/other/path`. We've been using the first of those three URLs as the base URL for the version when rendering it in diffs, but this change lets us use the last of those three URLs instead.

Fixes #600.