Show details of version URLs and redirects (#770)
If either version being shown on the page details screen had redirects or used URLs that didn’t exactly match the page’s current canonical URL, we now show a little ⚠️ warning message about it. Clicking it discloses more details, with a complete list of redirects and URLs for each version. The parts of the URL that differ (either from the page’s canonical URL or from the previous URL it was redirected from) are subtly highlighted.

There are three types of warning, each of which is a superset of the previous:
1. "Captured from a different URL" (There were no redirects, but the URL was not an *exact* match to the page's, either because it was similar enough or because the page's URL has changed over time.)
2. "Captured URL Redirected Somewhere Else"
3. "Versions come from different URLs"

Fixes #509.