Show monitored status, HTTP status, and tags
Add the page's monitored status and HTTP status to the list view.
Add the page's monitored status and HTTP status and tags to the page details view.

The DOM structure on page details has gotten a bit messy, and it would be good to clean up. I don't think that's worth blocking this for, though.

Fixes #331.