Show monitored status, HTTP status, and tags (#771)
This adds the page’s monitored status (i.e. whether we are recording new versions) and the HTTP status to both the list view and the page details view. It also adds tags to the page details view.

The DOM structure on the page details view has gotten a bit messy, and it would be good to clean up, but that can be done in a follow-up.

Fixes #331.