Use new sample endpoint to load less data (#956)
In edgi-govdata-archiving/web-monitoring-db#946, we added a new API endpoint at `/api/v0/pages/:page_id/versions/sampled` that gets a "sampled" list of versions — that is, one version per sampling period (for now, the sampling period is always 1 day, so this is one version per day). This makes use of it.

The main idea here is to load data reasonably quickly — some pages (e.g. `https://epa.gov/`) have a *lot* of versions, and require hundreds of requests and many megabytes of data in order to load the details page. Loading a one-per-day sample makes pages much quicker to load, with fewer, smaller requests, but still giving people a complete-enough list of versions to select from.