Don't require logging in (#1037)
If the `ALLOW_PUBLIC_VIEW` config variable is set to `true`, the UI will no longer show a modal dialog requiring someone to be logged in before trying to browse any data. Having this work depends on the API actually allowing unauthenticated requests (at the time of writing, that's allowed in staging but not yet in production).

Fixes #1025.