Use correct name for media type "essence" (#1041)
Back when I made our representation for media types, I was trying to figure out the right name of the `{type}/{subtype}` (without parameters) string form of the media type because I wasn't aware there *was* a correct name. I gave up and just called it `mediaType`. TIL it is properly referred to as the "essence": https://mimesniff.spec.whatwg.org/#understanding-mime-types