Sync scrolling in side-by-side diff (#1073)
Add an option (on by default) to synchronize scrolling in side-by-side diffs. Because the iframes for each view may have a different origin, this uses `postMessage()` for each frame to broadcast its position when scrolled, and the main app forwards those broadcasted messages to the other frame, which updates its scroll position.

This also has a crazy "landmark" system for tracking scroll position. When there are big changes to the content, matching parts of the page may be offset from each other in pixels. We want to keep matching parts of the page visually scrolled together, though, so this attempts to find unchanging landmark elements that are the same on both sides of the diff and sets scroll positions relative to those, rather than setting them in absolute pixels.