Replace Enzyme with React Testing Library (#1078)
This app is trapped on v16 of React (current is v19) because of some old tooling. One roadbloack is our usage of Enzyme for testing, which never shipped a working version for React v17. The new replacement is React Testing Library, which this switches us to. It has pretty good community support and has survived for 3 major React releases, so I think it’s relatively safe.

Note this uses React Testing Library v12, which is nowhere near current, but is the latest that is compatible with React 16. Newer versions require React 18, so we’ll need to upgrade this along with React (if and when we do that).

This also adds jest-dom, which is just a nicer set of DOM-specific assertions for Jest.