Uncontroversial Dependency Updates
Since this project is semi-maintained again, this updates all our dependencies that reasonably can be without *too* much change to our code or deep inspection of the impacts. This is basically just stuff where there are only minor updates, or where the breaking changes are in features we don't use or requires Node.js versions that are safe for us (i.e. >= 18.x). This still leaves a lot of things out-of-date that require more work.