Switch to modern React contexts (#1083)
We use React's context API to pass around instances of the API services for different components to use, but depended on React's original context API, which was deprecated 7 years ago. This switches to the new API. It's a little hacky -- the new API only allows class components to use a single context, when it would be nice to have separate contexts for the DB and local APIs. But we frequently use them together, so I just a single context that contains both.