Drop babel-plugin-react-css-modules (#1087)
This drops the babel-plugin-react-css-modules package and instead uses css-loader directly. The Babel plugin provided some nice, convenient syntax, but hasn’t been well maintained. Its approach also causes frequent breakage because it needs to keep in sync with project config and implementation details of css-loader. There is a fork that is well maintained, but I decided not to switch to it because the fundamental issues (and regular breakage) remain.

This also changes test snapshots slightly — with the Babel plugin, the selector re-naming is done by Babel and happens during tests. Without the Babel plugin, re-naming is done by Webpack, and does not actually run during Jest tests. So instead we have a little mock that just gives you back the selector name you asked for as a string.

Fixes #581.