Don't re-render diff when changing scroll sync
When changing whether scrolling was sync'd, it would cause the diff views to re-render and reset their scrolling to the top (so you couldn't, for example, scroll to a position, unsync, and then scroll more on one side). This was happening because changing the setting (re-)creates a new set of transforms, which are different props for the `SandboxedHtmlView`, which then re-sets the `srcdoc` property on its `iframe`. This solves the issue by only re-setting `srcdoc` if we are actually changing it. (Ideally we'd also do some caching so that creating a transform with the same arguments returns an idental transformer function, but that's a whole other ball of wax with memory management, and is not the only way we can cause this sort of problem. This fix is dumber, but probably more resilient.)