Remove node-fetch and fetch-mock (#1088)
We do not appear to actually use either of these packages (we've only been using `fetch` in Node since v18, which has it built-in, and there are no references to fetch-mock anywhere in tests), so this drops them.

This also updates some transitive dependencies to resolve a few security warnings.