Replace extract-loader
The old extract-loader Webpack plugin is no longer maintained and is the source of all our remaining dependency security warnings. Looking at its issue tracker, it seems like there are also a number of edge-cases where it does not work correctly with Webpack 5 that have been slowly piling up (we haven't hit those yet, but who knows when that might happen). This switches to the currently recommended by Webpack mini-css-extract-plugin.