Remove react-dates
The react-dates package provided a nice (although maybe a little over-sized) datepicker. However, it doesn't support React 17 or newer (that is, no version of React released in the last 5 years!) and doesn't appear that it is going to any time in the near future. The feature it powers isn't really used much, so I've opted to simply replace it with two browser-native date inputs rather than switch to some other more maintained range picker component.

This replaces the functionality, but needs some more styling and structuring work.