Remove react-dates, use Luxon for time (#1091)
React-dates is not compatible with React 17+, and is preventing us from upgrading lots of other tools and packages to currently supported versions. Its author stated it won't be upgraded until Enzyme is, and my current belief is that Enzyme is effectively dead (same author). This removes it so we can move forward.

This also switches us from Moment to Luxon for time handling. Moment has been deprecated for years, and we only remained on it because react-dates depended on it.

Part of the modernization work in #1082.