Discourage robots from indexing this app (#1096)
Per a discussion with the team, it's probably not good or useful if random comparison views from this app show up on Google, or if Internet Archive archives our view of their archives. (Context: since we show a lot of pages that no longer exist, Google has started pointing to us for some things! But we're not really an optimal view of a historical capture -- people should go to the Internet Archive for that.)