Fix `diffTypesFor()` when using a file extension (#1097)
The `diffTypesFor()` function is intended to allow you to pass in a MediaType object, media type string, or a file extension (e.g. `".pdf"`). However, it turns out the file extension is broken (I guess we have pretty much stopped using file extensions since we improved media type support way back when...). This fixes the issue and adds a test.

This bug was originally uncovered by @BeckettFrey in #1095.