Upgrade ESLint to 9.35.0 (#1102)
This upgrades ESLint to the 9.x release line, which uses an entirely different config format. This also brings a lot of the ESLint config and tooling up-to-date, and it flagged a bunch of things it was not before, so there is a bunch of syntax cleanup that came along for the ride here.

Part of #1082.