Add cooldown period for Dependabot updates (#1107)
Thinking about the big supply chain issue on Monday and the mess of Luxon patches yesterday (see #1105), I think it's good to have a little bit of a cooldown period for dependencies. It’s a little longer for NPM packages, since they tend to be a little less stable than many other types in my experience.