Feature/handle non renderable content 1067 (#1095)
This adds a new diff type that renders a "file preview," which lists some basic metadata about the file and a button to view/download the raw content. For most media types, this replaces the "raw content" view, which just puts the content in an iframe, and often caused the browser to download the file instead of showing you anything meaningful (for types the browser can’t render natively).

Fixes #1067.